feat: added mutations to insert rates
