+++ src/hooks/useErrors.js
@@ -1,5 +1,5 @@
 import { useMutation } from '@apollo/client'
-import { INSERT_ERRORS_MUTATION } from '../services/GraphQl/mutations/errors'
+import { INSERT_ERRORS_MUTATION } from '../services/GraphQl'
 
 export function InsertErrors() {
   const [insertErrors, { loading, error, data }] = useMutation(INSERT_ERRORS_MUTATION)
