+++ src/services/GraphQl/subscription/index.js
@@ -1,3 +1,3 @@
-export * from './package'
+export * from './package.subscriptions'
